app/requirements.txt
fastapi
uvicorn
jinja2
aiohttp
beautifulsoup4
python-multipart
python-dotenv